feat(cal-ui-p4): move time/andén picker into package (generic)
The departure-time + andén availability picker (TimeSlotAssignment) is now the
domain-agnostic component in @microboxlabs/miot-calendar-ui: options are
host-computed and all labels resolve through the CalendarHost i18n seam. A
back-compat shim at the old sidebar-tabs path forwards to it.
